Release: v2.0.0

+++ TelegramGroupsAdmin.Core/JobPayloads/FetchUserPhotoPayload.cs
@@ -5,6 +5,6 @@ namespace TelegramGroupsAdmin.Core.JobPayloads;
 /// Downloads and caches user profile photo, then updates message record
 /// </summary>
 public record FetchUserPhotoPayload(
-    long MessageId,
+    int MessageId,
     long UserId
 );
